# src/controllers/ProductController.js
const { CONFIG_MESSAGE_ERRORS } = require("../configs");
const ProductService = require("../services/ProductService");
const { validateRequiredInput, validateDiscountDate } = require("../utils");

const createProduct = async (req, res) => {
	try {
		const requiredFields = validateRequiredInput(req.body, [
			"name",
			"image",
			"type",
			"countInStock",
			"price",
		]);
		const discountValidation = validateDiscountDate(
			req.body.discount,
			new Date(req?.body?.discountStartDate),
			new Date(req?.body?.discountEndDate)
		);

		if (requiredFields?.length) {
			return res.status(CONFIG_MESSAGE_ERRORS.INVALID.status).json({
				status: "Error",
				typeError: CONFIG_MESSAGE_ERRORS.INVALID.type,
				message: `The field ${requiredFields.join(", ")} is required`,
				data: null,
			});
		}
		if (!discountValidation.isValid) {
			return res.status(CONFIG_MESSAGE_ERRORS.INVALID.status).json({
				status: "Error",
				message: discountValidation.error,
				typeError: CONFIG_MESSAGE_ERRORS.INVALID.type,
				data: null,
			});
		}
		const response = await ProductService.createProduct(req.body);
		const { data, status, typeError, message, statusMessage } = response;
		return res.status(status).json({
			typeError,
			data,
			message,
			status: statusMessage,
		});
	} catch (e) {
		return res.status(CONFIG_MESSAGE_ERRORS.INTERNAL_ERROR.status).json({
			message: "Internal Server Error",
			data: null,
			status: "Error",
			typeError: CONFIG_MESSAGE_ERRORS.INTERNAL_ERROR.type,
		});
	}
};

const updateProduct = async (req, res) => {
	try {
		const productId = req.params.id;
		const discountValidation = validateDiscountDate(
			req.body.discount,
			new Date(req.body.discountStartDate),
			new Date(req.body.discountEndDate)
		);

		const requiredFields = validateRequiredInput(req.body, [
			"name",
			"image",
			"type",
			"countInStock",
			"price",
		]);

		if (!productId) {
			return res.status(CONFIG_MESSAGE_ERRORS.INVALID.status).json({
				status: "Error",
				typeError: CONFIG_MESSAGE_ERRORS.INVALID.type,
				message: `The field productId is required`,
			});
		}
		if (requiredFields?.length) {
			return res.status(CONFIG_MESSAGE_ERRORS.INVALID.status).json({
				status: "Error",
				typeError: CONFIG_MESSAGE_ERRORS.INVALID.type,
				message: `The field ${requiredFields.join(", ")} is required`,
				data: null,
			});
		}
		if (!discountValidation.isValid) {
			return res.status(CONFIG_MESSAGE_ERRORS.INVALID.status).json({
				status: "Error",
				message: discountValidation.error,
				typeError: CONFIG_MESSAGE_ERRORS.INVALID.type,
				data: null,
			});
		}
		const response = await ProductService.updateProduct(productId, req.body);
		const { data, status, typeError, message, statusMessage } = response;
		return res.status(status).json({
			typeError,
			data,
			message,
			status: statusMessage,
		});
	} catch (e) {
		return res.status(CONFIG_MESSAGE_ERRORS.INTERNAL_ERROR.status).json({
			message: "Internal Server Error",
			data: null,
			status: "Error",
			typeError: CONFIG_MESSAGE_ERRORS.INTERNAL_ERROR.type,
		});
	}
};

const getDetailsProduct = async (req, res) => {
	try {
		const productId = req.params.id;
		if (!productId) {
			return res.status(CONFIG_MESSAGE_ERRORS.INVALID.status).json({
				status: "Error",
				typeError: CONFIG_MESSAGE_ERRORS.INVALID.type,
				message: `The field productId is required`,
			});
		}
		const response = await ProductService.getDetailsProduct(productId);
		const { data, status, typeError, message, statusMessage } = response;
		return res.status(status).json({
			typeError,
			data,
			message,
			status: statusMessage,
		});
	} catch (e) {
		console.log("products", { e });
		return res.status(CONFIG_MESSAGE_ERRORS.INTERNAL_ERROR.status).json({
			message: "Internal Server Error",
			data: null,
			status: "Error",
			typeError: CONFIG_MESSAGE_ERRORS.INTERNAL_ERROR.type,
		});
	}
};

const getDetailsProductPublic = async (req, res) => {
	try {
		const productId = req.params.id;
		const userId = req?.userId;
		const params = req.query;

		if (!productId) {
			return res.status(CONFIG_MESSAGE_ERRORS.INVALID.status).json({
				status: "Error",
				typeError: CONFIG_MESSAGE_ERRORS.INVALID.type,
				message: `The field productId is required`,
			});
		}
		const response = await ProductService.getDetailsProductPublic(
			productId,
			userId,
			params
		);
		const { data, status, typeError, message, statusMessage } = response;
		return res.status(status).json({
			typeError,
			data,
			message,
			status: statusMessage,
		});
	} catch (e) {
		return res.status(CONFIG_MESSAGE_ERRORS.INTERNAL_ERROR.status).json({
			message: "Internal Server Error",
			data: null,
			status: "Error",
			typeError: CONFIG_MESSAGE_ERRORS.INTERNAL_ERROR.type,
		});
	}
};

const getDetailsProductPublicBySlug = async (req, res) => {
	try {
		const slugId = req.params.slug;
		const userId = req?.userId;
		const params = req.query;

		if (!slugId) {
			return res.status(CONFIG_MESSAGE_ERRORS.INVALID.status).json({
				status: "Error",
				typeError: CONFIG_MESSAGE_ERRORS.INVALID.type,
				message: `The field slugId is required`,
			});
		}
		const response = await ProductService.getDetailsProductPublicBySlug(
			slugId,
			userId,
			params
		);
		const { data, status, typeError, message, statusMessage } = response;
		return res.status(status).json({
			typeError,
			data,
			message,
			status: statusMessage,
		});
	} catch (e) {
		return res.status(CONFIG_MESSAGE_ERRORS.INTERNAL_ERROR.status).json({
			message: "Internal Server Error",
			data: null,
			status: "Error",
			typeError: CONFIG_MESSAGE_ERRORS.INTERNAL_ERROR.type,
		});
	}
};

const deleteProduct = async (req, res) => {
	try {
		const productId = req.params.id;
		if (!productId) {
			return res.status(CONFIG_MESSAGE_ERRORS.INVALID.status).json({
				status: "Error",
				typeError: CONFIG_MESSAGE_ERRORS.INVALID.type,
				message: `The field productId is required`,
			});
		}
		const response = await ProductService.deleteProduct(productId);
		const { data, status, typeError, message, statusMessage } = response;
		return res.status(status).json({
			typeError,
			data,
			message,
			status: statusMessage,
		});
	} catch (e) {
		return res.status(CONFIG_MESSAGE_ERRORS.INTERNAL_ERROR.status).json({
			message: "Internal Server Error",
			data: null,
			status: "Error",
			typeError: CONFIG_MESSAGE_ERRORS.INTERNAL_ERROR.type,
		});
	}
};

const deleteMany = async (req, res) => {
	try {
		const ids = req.body.productIds;
		if (!ids || !ids.length) {
			return res.status(CONFIG_MESSAGE_ERRORS.INVALID.status).json({
				status: "Error",
				typeError: CONFIG_MESSAGE_ERRORS.INVALID.type,
				message: `The field productIds is required`,
			});
		}
		const response = await ProductService.deleteManyProduct(ids);
		const { data, status, typeError, message, statusMessage } = response;
		return res.status(status).json({
			typeError,
			data,
			message,
			status: statusMessage,
		});
	} catch (e) {
		return res.status(CONFIG_MESSAGE_ERRORS.INTERNAL_ERROR.status).json({
			message: "Internal Server Error",
			data: null,
			status: "Error",
			typeError: CONFIG_MESSAGE_ERRORS.INTERNAL_ERROR.type,
		});
	}
};

const getAllProduct = async (req, res) => {
	try {
		const params = req.query;
		const response = await ProductService.getAllProduct(params);
		const { data, status, typeError, message, statusMessage } = response;
		return res.status(status).json({
			typeError,
			data,
			message,
			status: statusMessage,
		});
	} catch (e) {
		return res.status(CONFIG_MESSAGE_ERRORS.INTERNAL_ERROR.status).json({
			message: "Internal Server Error",
			data: null,
			status: "Error",
			typeError: CONFIG_MESSAGE_ERRORS.INTERNAL_ERROR.type,
		});
	}
};

const getAllProductPublic = async (req, res) => {
	try {
		const params = req.query;
		const response = await ProductService.getAllProductPublic(params);
		const { data, status, typeError, message, statusMessage } = response;
		return res.status(status).json({
			typeError,
			data,
			message,
			status: statusMessage,
		});
	} catch (e) {
		return res.status(CONFIG_MESSAGE_ERRORS.INTERNAL_ERROR.status).json({
			message: "Internal Server Error",
			data: null,
			status: "Error",
			typeError: CONFIG_MESSAGE_ERRORS.INTERNAL_ERROR.type,
		});
	}
};

const likeProduct = async (req, res) => {
	try {
		const { productId } = req.body;
		const userId = req.userId;

		if (!productId) {
			return res.status(CONFIG_MESSAGE_ERRORS.INVALID.status).json({
				status: "Error",
				typeError: CONFIG_MESSAGE_ERRORS.INVALID.type,
				message: `The field productId is required`,
				data: null,
			});
		}
		const response = await ProductService.likeProduct(productId, userId);
		const { data, status, typeError, message, statusMessage } = response;
		return res.status(status).json({
			typeError,
			data,
			message,
			status: statusMessage,
		});
	} catch (e) {
		return res.status(CONFIG_MESSAGE_ERRORS.INTERNAL_ERROR.status).json({
			message: "Internal Server Error",
			data: null,
			status: "Error",
			typeError: CONFIG_MESSAGE_ERRORS.INTERNAL_ERROR.type,
		});
	}
};

const unlikeProduct = async (req, res) => {
	try {
		const { productId } = req.body;
		const userId = req.userId;

		if (!productId) {
			return res.status(CONFIG_MESSAGE_ERRORS.INVALID.status).json({
				status: "Error",
				typeError: CONFIG_MESSAGE_ERRORS.INVALID.type,
				message: `The field productId is required`,
				data: null,
			});
		}
		const response = await ProductService.unlikeProduct(productId, userId);
		const { data, status, typeError, message, statusMessage } = response;
		return res.status(status).json({
			typeError,
			data,
			message,
			status: statusMessage,
		});
	} catch (e) {
		return res.status(CONFIG_MESSAGE_ERRORS.INTERNAL_ERROR.status).json({
			message: "Internal Server Error",
			data: null,
			status: "Error",
			typeError: CONFIG_MESSAGE_ERRORS.INTERNAL_ERROR.type,
		});
	}
};

const getAllProductLiked = async (req, res) => {
	try {
		const userId = req.userId;
		const params = req.query;

		const response = await ProductService.getAllProductLiked(userId, params);
		const { data, status, typeError, message, statusMessage } = response;
		return res.status(status).json({
			typeError,
			data,
			message,
			status: statusMessage,
		});
	} catch (e) {
		return res.status(CONFIG_MESSAGE_ERRORS.INTERNAL_ERROR.status).json({
			typeError: "Internal Server Error",
			data: null,
			status: "Error",
		});
	}
};

const getAllProductViewed = async (req, res) => {
	try {
		const userId = req.userId;
		const params = req.query;
		const response = await ProductService.getAllProductViewed(userId, params);
		const { data, status, typeError, message, statusMessage } = response;
		return res.status(status).json({
			typeError,
			data,
			message,
			status: statusMessage,
		});
	} catch (e) {
		return res.status(CONFIG_MESSAGE_ERRORS.INTERNAL_ERROR.status).json({
			typeError: "Internal Server Error",
			data: null,
			status: "Error",
		});
	}
};

const getListRelatedProductBySlug = async (req, res) => {
	try {
		const params = req.query;
		if (!params.slug) {
			return res.status(CONFIG_MESSAGE_ERRORS.INVALID.status).json({
				status: "Error",
				typeError: CONFIG_MESSAGE_ERRORS.INVALID.type,
				message: `The field slug is required`,
			});
		}
		const response = await ProductService.getListRelatedProductBySlug(params);
		const { data, status, typeError, message, statusMessage } = response;
		return res.status(status).json({
			typeError,
			data,
			message,
			status: statusMessage,
		});
	} catch (e) {
		return res.status(CONFIG_MESSAGE_ERRORS.INTERNAL_ERROR.status).json({
			message: "Internal Server Error",
			data: null,
			status: "Error",
			typeError: CONFIG_MESSAGE_ERRORS.INTERNAL_ERROR.type,
		});
	}
};

module.exports = {
	createProduct,
	updateProduct,
	getDetailsProduct,
	deleteProduct,
	getAllProduct,
	deleteMany,
	likeProduct,
	unlikeProduct,
	getDetailsProductPublic,
	getAllProductPublic,
	getAllProductViewed,
	getAllProductLiked,
	getDetailsProductPublicBySlug,
	getListRelatedProductBySlug,
};